Replace existing build with new logic
