+++ .eslintignore
@@ -1,7 +1,6 @@
 node_modules
 /examples
 /build
-/build-intermediate
 /docs
 /esnext
 /types
