+++ .stylelintignore
@@ -1,6 +1,5 @@
 node_modules
 /build
-/build-intermediate
 /docs
 /esnext
 /types
