+++ .vscode/settings.json
@@ -13,7 +13,6 @@
     "**/.svn": true,
     "**/node_modules": true,
     ".{dev,shadowenv.d}": true,
-    "build-intermediate/": true,
     "build/": true,
     "scripts/analyze-custom-properties/dist/": true,
     "docs/": true,
@@ -31,7 +30,6 @@
     "**/bower_components": true,
     "**/node_modules": true,
     ".{dev,shadowenv.d}": true,
-    "build-intermediate/": true,
     "build/": true,
     "docs/": true,
     "esnext/": true,
