--- scripts/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "exclude": [
-    "../src/**/*.test.ts",
-    "../src/**/*.test.tsx",
-    "../tests",
-    "../playground"
-  ]
-}
